# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 语赞心声 API
  version: 0.1.0
  description: Contract baseline for the first real teaching loop.
  license:
    name: Proprietary
    url: https://example.edu/license
servers:
  - url: /api/v1
    description: Current deployment origin
...
paths:
...
  /schools/{schoolId}/pilot/overview:
    get:
      summary: 获取试点运行概览
      tags: [Pilot]
      operationId: getPilotOverview
      parameters:
        - $ref: "#/components/parameters/SchoolId"
        - name: window
          in: query
          schema:
            type: string
            enum: [24h, 7d]
            default: 24h
      responses:
        "200":
          description: 试点运行概览
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /schools/{schoolId}/pilot/feedback:
    post:
      summary: 提交试点产品反馈
      tags: [Pilot]
      operationId: createPilotFeedback
      parameters:
        - $ref: "#/components/parameters/SchoolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePilotFeedbackRequest"
      responses:
        "201":
          description: 反馈已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotFeedbackCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      summary: 管理员查看试点反馈
      tags: [Pilot]
      operationId: listPilotFeedback
      parameters:
        - $ref: "#/components/parameters/SchoolId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        - name: category
          in: query
          schema:
            {
              type: string,
              enum:
                [
                  CONTENT,
                  MEDIA,
                  RECORDING,
                  SCORING,
                  USABILITY,
                  TECHNICAL,
                  OTHER,
                ],
            }
      responses:
        "200":
          description: 反馈列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotFeedbackListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /schools/{schoolId}/pilot/feedback/mine:
    get:
      summary: 查看自己的试点反馈
      tags: [Pilot]
      operationId: listMyPilotFeedback
      parameters:
        - $ref: "#/components/parameters/SchoolId"
      responses:
        "200":
          description: 自己提交的反馈
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotFeedbackListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /schools/{schoolId}/pilot/feedback/{feedbackId}/status:
    patch:
      summary: 更新试点反馈处理状态
      tags: [Pilot]
      operationId: updatePilotFeedbackStatus
      parameters:
        - $ref: "#/components/parameters/SchoolId"
        - name: feedbackId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePilotFeedbackStatusRequest"
      responses:
        "200":
          description: 状态已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotFeedbackStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SchoolId:
      in: path
      name: schoolId
      required: true
      description: 学校（租户）标识
      schema:
        type: string
        format: uuid
        example: "00000000-0000-0000-0000-000000000000"
    Limit:
      in: query
      name: limit
      description: 分页返回数量上限
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      in: query
      name: cursor
      description: 游标， opaque 字符串
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 请求参数无效或格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INVALID_REQUEST
              message: 请求参数无效
              details: {}
              requestId: "00000000-0000-0000-0000-000000000002"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but outside allowed resource scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource does not exist, has been soft-deleted, or is not visible in the current school scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    CreatePilotFeedbackRequest:
      type: object
      required: [category, message]
      additionalProperties: false
      properties:
        category:
          {
            type: string,
            enum:
              [CONTENT, MEDIA, RECORDING, SCORING, USABILITY, TECHNICAL, OTHER],
          }
        message: { type: string, minLength: 5, maxLength: 1000 }
        sessionId: { type: [string, "null"], format: uuid }
        assessmentItemId: { type: [string, "null"], format: uuid }
        currentPath: { type: [string, "null"], maxLength: 500 }
    UpdatePilotFeedbackStatusRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status: { type: string, enum: [ACKNOWLEDGED, RESOLVED] }
        resolutionNote:
          { type: [string, "null"], minLength: 1, maxLength: 1000 }
    PilotFeedbackCreateResponse:
      type: object
      required: [feedbackId, status, createdAt]
      properties:
        feedbackId: { type: string, format: uuid }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        createdAt: { type: string, format: date-time }
    PilotFeedbackListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PilotFeedbackItem"
        nextCursor: { type: [string, "null"] }
    PilotFeedbackItem:
      type: object
      required: [id, category, status, message, createdAt]
      properties:
        id: { type: string, format: uuid }
        category:
          {
            type: string,
            enum:
              [CONTENT, MEDIA, RECORDING, SCORING, USABILITY, TECHNICAL, OTHER],
          }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        message: { type: string }
        pageContext: { type: [string, "null"] }
        context: { type: [object, "null"], additionalProperties: true }
        reporter: { type: [object, "null"], additionalProperties: true }
        resolutionNote: { type: [string, "null"] }
        resolvedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    PilotFeedbackStatusResponse:
      type: object
      required: [id, status, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [ACKNOWLEDGED, RESOLVED] }
        resolutionNote: { type: [string, "null"] }
        resolvedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    PilotOverviewResponse:
      type: object
      required:
        [
          window,
          generatedAt,
          overallState,
          dependencies,
          formalSessions,
          standardSessions,
          processingBacklog,
          teacherReviewBacklog,
          remediation,
          speech,
          feedback,
          warnings,
        ]
      properties:
        window: { type: string, enum: [24h, 7d] }
        generatedAt: { type: string, format: date-time }
        overallState: { type: string, enum: [HEALTHY, ATTENTION, DEGRADED] }
        dependencies: { type: object, additionalProperties: true }
        formalSessions: { type: object, additionalProperties: true }
        standardSessions: { type: object, additionalProperties: true }
        processingBacklog: { type: object, additionalProperties: true }
        teacherReviewBacklog: { type: object, additionalProperties: true }
        remediation: { type: object, additionalProperties: true }
        speech: { type: object, additionalProperties: true }
        feedback: { type: object, additionalProperties: true }
        warnings: { type: array, items: { type: string } }
...
    ErrorBody:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]*$"
          minLength: 1
        message:
          type: string
        details:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
              - type: "null"
              - type: array
                items:
                  type: string
        requestId:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
      example:
        error:
          code: ASSIGNMENT_NOT_OPEN
          message: 当前任务尚未开放
          details:
            assignmentId: "11111111-1111-1111-1111-111111111111"
          requestId: "00000000-0000-0000-0000-000000000003"
...
security:
  - bearerAuth: []